Move NonceInfo, NonceFull and NoncePartial out of bank.rs
