+++ runtime/src/accounts.rs
@@ -13,8 +13,9 @@ use {
         },
         accounts_update_notifier_interface::AccountsUpdateNotifier,
         ancestors::Ancestors,
-        bank::{Bank, NonceFull, NonceInfo, TransactionCheckResult, TransactionExecutionResult},
+        bank::{Bank, TransactionCheckResult, TransactionExecutionResult},
         blockhash_queue::BlockhashQueue,
+        nonce_info::{NonceFull, NonceInfo},
         rent_collector::RentCollector,
         rent_debits::RentDebits,
         storable_accounts::StorableAccounts,
